added pageview-dynamic that was missing
