+++ docs/src/components/pageview.tsx
@@ -1,26 +1,9 @@
 'use client';
 
-import { usePathname, useSearchParams } from "next/navigation";
-import { usePostHog } from 'posthog-js/react';
-import { useEffect } from "react";
+import dynamic from "next/dynamic";
 
-export default function PostHogPageView(): null {
-  const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const posthog = usePostHog();
+const PageView = dynamic(() => import('./pageview-dynamic'), {
+  ssr: false,
+});
 
-  useEffect(() => {
-    let url = window.origin + pathname;
-    if (searchParams.toString()) {
-      url = url + `?${searchParams.toString()}`;
-    }
-    posthog.capture(
-      '$pageview',
-      {
-        '$current_url': url,
-      }
-    );
-  }, [pathname, searchParams, posthog]);
-
-  return null;
-}
+export default PageView;
